Rename `Lambda.Lifecycle` to `LambdaRuntime` (#232)
